Improvements in performanceAnalysis
