--- performanceAnalysis/sorting Algorithms/performanceMeasure.java
@@ -1,8 +1,0 @@
-/* 
-    Contributions Required!
-    Please Read the documentation and go through the repository
-    before beginning to contribute.
-
-    This is where Sorting algorithms will be tested against a
-    randomly generated mutual data set.
-*/
